+++ src/main/java/org/liubov/ai_aggregator/model/ImageRequest.java
@@ -15,11 +15,14 @@ public class ImageRequest {
     @Id
     @GeneratedValue(strategy= GenerationType.IDENTITY)
     @Setter(value = AccessLevel.PRIVATE)
-    private Integer id;
+    private Long id;
+
     @Column(nullable = false)
-    private Integer imageDialogId;
+    private Long imageDialogId;
+
     @Column(nullable = false)
     private String text;
+
     @CreationTimestamp
     private LocalDateTime createdOn;
 }
