+++ src/main/java/org/liubov/ai_aggregator/repository/ChatModelRepository.java
@@ -3,5 +3,5 @@ package org.liubov.ai_aggregator.repository;
 import org.liubov.ai_aggregator.model.ChatModel;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ChatModelRepository extends JpaRepository<ChatModel, Integer> {
+public interface ChatModelRepository extends JpaRepository<ChatModel, Long> {
 }
