+++ src/main/java/org/liubov/ai_aggregator/repository/ImageDialogRepository.java
@@ -3,5 +3,5 @@ package org.liubov.ai_aggregator.repository;
 import org.liubov.ai_aggregator.model.ImageDialog;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ImageDialogRepository extends JpaRepository<ImageDialog, Integer> {
+public interface ImageDialogRepository extends JpaRepository<ImageDialog, Long> {
 }
