+++ src/main/java/org/liubov/ai_aggregator/repository/ImageModelRepository.java
@@ -3,5 +3,5 @@ package org.liubov.ai_aggregator.repository;
 import org.liubov.ai_aggregator.model.ImageModel;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ImageModelRepository extends JpaRepository<ImageModel, Integer> {
+public interface ImageModelRepository extends JpaRepository<ImageModel, Long> {
 }
