+++ src/main/java/org/liubov/ai_aggregator/repository/ImageRequestRepository.java
@@ -3,5 +3,5 @@ package org.liubov.ai_aggregator.repository;
 import org.liubov.ai_aggregator.model.ImageRequest;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ImageRequestRepository extends JpaRepository<ImageRequest, Integer> {
+public interface ImageRequestRepository extends JpaRepository<ImageRequest, Long> {
 }
