+++ src/main/java/org/liubov/ai_aggregator/repository/ImageResponseRepository.java
@@ -3,5 +3,5 @@ package org.liubov.ai_aggregator.repository;
 import org.liubov.ai_aggregator.model.ImageResponse;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ImageResponseRepository extends JpaRepository<ImageResponse, Integer> {
+public interface ImageResponseRepository extends JpaRepository<ImageResponse, Long> {
 }
